feat: initial repository classes for api basis
Model refactorings: (unidirectional) relationships and renaming to MessageX
