+++ src/main/java/org/eclipse/tractusx/puris/backend/common/api/domain/model/MessageContentError.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2023 Volkswagen AG
+ * Copyright (c) 2023 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e.V. (represented by Fraunhofer ISST)
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.eclipse.tractusx.puris.backend.common.api.domain.model;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+/**
+ * Implementation of a {@link MessageContent} providing generic error messages.
+ */
+@Entity
+@Table(name = "MessageContentError")
+@DiscriminatorValue("ERROR")
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+public class MessageContentError extends MessageContent {
+
+    /**
+     * identifies a business object of a specific api that refers to the error.
+     */
+    @OneToOne(cascade = CascadeType.ALL, orphanRemoval = true)
+    @JoinColumn(name = "reference_identification_uuid")
+    private ReferenceIdentification referenceIdentification;
+
+    /**
+     * error code specifying the error.
+     */
+    @NotNull
+    private String error;
+
+    /**
+     * the actual error message providing further detail.
+     */
+    @NotNull
+    private String message;
+}
