+++ src/main/java/org/eclipse/tractusx/puris/backend/common/api/domain/model/ReferenceIdentification.java
@@ -18,54 +18,33 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.eclipse.tractusx.puris.backend.common.api.domain;
+package org.eclipse.tractusx.puris.backend.common.api.domain.model;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.eclipse.tractusx.puris.backend.common.api.domain.datatype.DT_RequestStateEnum;
 
-import java.util.List;
 import java.util.UUID;
 
 /**
- * This Response represents the message received via a Response API.
- * <p>
- * This Response may not be confused with an HTTP response.
- * Both, the Response and the Request, are called (api) request.
+ * Allows different Identification Patterns for the api {@link MessageContent}
  */
 @Entity
-@Table(name = "Response")
+@Table(name = "ReferenceIdentification")
+@DiscriminatorColumn(name = "reference_identification_type")
 @Getter
 @Setter
 @ToString
 @NoArgsConstructor
-public class Response {
+public abstract class ReferenceIdentification {
 
     /**
-     * This ID prevents the application from collision with external IDs, because the partner
-     * creates the request when performing a Request API call.
+     * Technical identifier for a Reference Identification.
      */
     @Id
     @GeneratedValue
-    private UUID internalRequestUuid;
+    private UUID uuid;
 
-    /**
-     * State of the request.
-     *
-     * @See DT_RequestStateEnum
-     */
-    @NotNull
-    private DT_RequestStateEnum state;
-
-    @OneToMany
-    @JoinColumn(name = "message_content_id")
-    @ToString.Exclude
-    /**
-     * Actual content of the request (or response) message.
-     */
-    private List<MessageContent> messageContents;
 }
